+++ src/utils/index.ts
@@ -5,7 +5,8 @@ export {
   createStorageHashKey,
   filterByIncludes,
   getStatusButtonElem,
-  handleKeyCommand
+  handleKeyCommand,
+  removeBackslashes,
 } from './helpers';
 
 export { USER_PERMS } from './constants';